utils_for_tokens.c OK
