+++ utils_for_tokens.c
@@ -10,65 +10,8 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-// Norm: Too many functions in the file
-
 #include "minishell.h"
 
-// NOTE was lst_new_token
-t_token	*new_node(char *str, int type, int status)
-{
-	t_token	*new_node;
-
-	new_node = malloc(sizeof(t_token));
-	if (!new_node)
-		return (NULL);
-	new_node->str = str;
-	new_node->str_backup = ft_strdup(str);
-	new_node->var_exists = false;
-	new_node->type = type;
-	new_node->status = status;
-	new_node->prev = NULL;
-	new_node->next = NULL;
-	return (new_node);
-}
-
-// NOTE was lst_add_back_token
-void	add_node(t_token **list, t_token *new_node)
-{
-	t_token	*start;
-
-	start = *list;
-	if (start == NULL)
-	{
-		*list = new_node;
-		return ;
-	}
-	if (list && *list && new_node)
-	{
-		while (start->next != NULL)
-		{
-			start = start->next;
-		}
-		start->next = new_node;
-		new_node->prev = start;
-	}
-}
-
-void	delete_node(t_token *node, void (*del)(void *))
-{
-	if (del && node && node->str)
-	{
-		(*del)(node->str);
-		node->str = NULL;
-	}
-	if (node->prev)
-		node->prev->next = node->next;
-	if (node->next)
-		node->next->prev = node->prev;
-	//free_node(node);
-}
-
-// NOTE was save_sepatator
 int	add_separator(t_token **token, char *str, int index, int type)
 {
 	int		i;
@@ -98,7 +41,6 @@ int	add_separator(t_token **token, char *str, int index, int type)
 	return (0);
 }
 
-// NOTE was save_word
 int	add_word(t_token **token, char *str, int index, int start)
 {
 	int		i;
@@ -115,28 +57,6 @@ int	add_word(t_token **token, char *str, int index, int start)
 	return (0);
 }
 
-// NOTE was is_separatetor
-int	which_separator(char *str, int i)
-{
-	if ((str[i] > 8 && str[i] < 14) || str[i] == 32)
-		return (SPACES);
-	else if (str[i] == '|')
-		return (PIPE);
-	else if (str[i] == '<' && str[i + 1] == '<')
-		return (HEREDOC);
-	else if (str[i] == '>' && str[i + 1] == '>')
-		return (APPEND);
-	else if (str[i] == '<')
-		return (INPUT);
-	else if (str[i] == '>')
-		return (TRUNC);
-	else if (str[i] == '\0')
-		return (END);
-	else
-		return (0);
-}
-
-// NOTE was save_word_or_sep
 int	add_word_or_sep(int *i, char *str, int start, t_global *global)
 {
 	int	type;
